tests/wpt/css-tests/css-ui-3_dev/html/caret-color-018.htm

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>CSS Basic User Interface Test: caret-color test animation</title>
<link href="mailto:rego@igalia.com" rel="author" title="Manuel Rego Casasnovas">
<link href="http://www.w3.org/TR/css3-ui/#caret-color" rel="help">
<link href="https://www.w3.org/TR/web-animations-1/#dom-animatable-animate" rel="help">
<link href="https://www.w3.org/TR/css3-color/#color0" rel="help">
<meta content="Test checks that caret-color is animatable as a color, and that the computed values during the animation are the expected ones." name="assert">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  textarea {
    caret-color: red;
  }
</style>
</head><body>
  <textarea id="textarea"></textarea>
  <div id="log"></div>

  <script>
    test(
      function(){
        var textarea = document.getElementById("textarea");
        assert_equals(getComputedStyle(textarea).caretColor, 'rgb(255, 0, 0)');

        var keyframes = [
          { caretColor: 'magenta' },
          { caretColor: 'lime' }
        ];
        var options = {
          duration: 10,
          fill: "forwards"
        };

        var player = textarea.animate(keyframes, options);
        player.pause();
        player.currentTime = 0;
        assert_equals(getComputedStyle(textarea).caretColor, 'rgb(255, 0, 255)');
        player.currentTime = 5;
        assert_equals(getComputedStyle(textarea).caretColor, 'rgb(128, 128, 128)');
        player.currentTime = 10;
        assert_equals(getComputedStyle(textarea).caretColor, 'rgb(0, 255, 0)');
      }, "The caret-color property is animatable");
</script>

</body></html>
